+++ src/ambition/version_details.cc
@@ -18,6 +18,12 @@ std::vector<VersionUpdates> versionDetails = {
     .identifier = _(IDENTIFIER_PREVIEW),
     .sections = {
       {
+        .heading = _(HEADING_ADDED),
+        .items = {
+          _("Translate War Machine identifiers (their type and level when they are in a list)."),
+        },
+      },
+      {
         .heading = _(HEADING_CHANGED),
         .items = {
           _("Music now plays on the tutorial and scenario selection screens."),